requirements.txt
pandas
requests
psycopg2-binary
matplotlib
seaborn